generalize ChangeFolderThread
